+++ src/detail/FrequencyMap.cpp
@@ -15,7 +15,10 @@ FrequencyMap::FrequencyMap(const std::vector<std::size_t> &frequencies)
 }
 
 std::size_t FrequencyMap::lookup(std::size_t x) const {
-  return std::upper_bound(begin(m_table), end(m_table), x) - begin(m_table);
+  // upper_bound(first, last, x) will always return in the range [first, last]
+  // Therefore upper_bound(begin(m_table), ...) - begin(m_table) will be >= 0
+  return static_cast<std::size_t>(
+    std::upper_bound(begin(m_table), end(m_table), x) - begin(m_table));
 }
 
 std::size_t FrequencyMap::sum() const { return m_sum; }
